Initial EF mocking
Tried a few different approaches but went for the one outlined in these
articles:

http://mylifeandcode.blogspot.co.uk/2012/10/unit-testing-with-mock-entity-framework.html
http://romiller.com/2012/02/14/testing-with-a-fake-dbcontext/

Suspect I may switch away from this approach to a repository pattern if
I ever decide to move to MongoDB.

Have a bad feeling that I've adopted an approach more suitable to an
older version of EF. Removing the virtual modifier from the DbSet (now
IDbSet) members of the context class (bARBieEntities) was a bit suspect.
However, this approach works and is quite clean.